# Adding @agray3's graph caching approach