@agray3 has [PR-8366](https://github.com/ggerganov/llama.cpp/pull/8366) open in mainline `llama.cpp` that appears to not meet the high standards of the `llama.cpp` maintainers. Me, being more pragmatic and less of a purist, would like to have these changes here as that way one avoids rebuilding the computation graph for every new token, a "feature" inherited from `llama.cpp` that I don't really like.

Here is what we get in performance improvement on CUDA (RTX-4080 with a Ryzen-7950X CPU)
| model             |       size |     params |          test |    t/s (main)    |       t/s (PR)   |  Speedup |
| ----------------- | ---------: | ---------: | ------------: | ---------------: | ---------------: | -------: |
| llama 8B Q4_0     |   4.33 GiB |     8.03 B |         tg128 |    123.55 ± 0.09 |    125.60 ± 0.11 |  1.017   |   
| llama 3B Q4_0     |   2.08 GiB |     3.61 B |         tg128 |    237.40 ± 1.03 |    244.19 ± 0.71 |  1.029   |   
| llama 1B Q4_0     | 933.24 MiB |     1.50 B |         tg128 |    519.27 ± 2.55 |    538.75 ± 2.32 |  1.038   |   
| llama 2-bpw TriLM |  45.84 MiB |    99.76 M |         tg128 |  1570.51 ± 49.67 |  1754.54 ± 64.75 |  1.117   | 

And here the performance improvement on Metal (M2-Max 30-core GPU, M2-Max CPU):
| model             |       size |          test |      t/s (main)  |    t/s (PR)      |  Speedup |
| ----------------- | ---------: | ------------: | ---------------: | ---------------: | -------: |
| llama 8B Q4_0     |   4.33 GiB |         tg128 |     59.38 ± 0.03 |     60.03 ± 0.03 |  1.011   |
| llama 3B Q4_0     |   2.08 GiB |         tg128 |    107.61 ± 0.55 |    108.74 ± 0.14 |  1.011   |
| llama 1B Q4_0     | 933.24 MiB |         tg128 |    225.92 ± 0.91 |    230.26 ± 0.76 |  1.019   |
| llama 2-bpw TriLM |  45.84 MiB |         tg128 |   520.46 ± 10.70 |    545.46 ± 7.33 |  1.048   |

The speedup obviously increases with decreasing model size as the time computing the graph becomes relatively shorter compared to the time taken building the graph. The speedup I observe is smaller compared to what @agray3 reports in  PR-8366. I guess, it is a matter of how fast the GPU is (where the graph is computed) relative to the CPU (where the graph is built).

GPU performance has not been a focus of this project. Still, how do we do relative to mainline llama.cpp after this PR? Using afd9909a (3942) from today, I get this for the RTX-4080

| model          |       size |          test |        t/s (mainline)|      t/s (PR)    |  Speedup  |
| ---------------| ---------: | ------------: | -------------------: | ---------------: | --------: |
| llama 8B Q4_0  |   4.33 GiB |         tg128 |        122.48 ± 0.10 |    125.60 ± 0.11 |  1.025    |   
| llama 3B Q4_0  |   2.08 GiB |         tg128 |        233.04 ± 0.66 |    244.19 ± 0.71 |  1.048    |   
| llama 1B Q4_0  | 933.24 MiB |         tg128 |        505.63 ± 1.23 |    538.75 ± 2.32 |  1.065    |   
  
and this for the M2-Max


 | model          |       size |          test |      t/s (mainline)  |     t/s (PR)     |  Speedup |
| ---------------| ---------: | ------------: | -------------------: | ---------------: | -------: |
| llama 8B Q4_0  |   4.33 GiB |         tg128 |         57.94 ± 0.32 |     60.03 ± 0.03 |  1.036   |
| llama 3B Q4_0  |   2.08 GiB |         tg128 |        103.67 ± 0.21 |    108.74 ± 0.14 |  1.049   |
| llama 1B Q4_0  | 933.24 MiB |         tg128 |        221.45 ± 1.31 |    230.26 ± 0.76 |  1.039   |


@agray3 Would you review the changes? Alternatively, if you prefer, we can close this PR and you can submit a PR yourself so this contribution is correctly associated with your name.