@ikawrakow 2024-10-19
Oh, btw, 

> @ikawrakow : check the "continuation" of this PR also :
> [ggerganov/llama.cpp#9017](https://github.com/ggerganov/llama.cpp/pull/9017)

Yes, I saw that. But the performance gain there is even less, so not sure if I want to add it.